refactor(pipeline): break 3 import cycles (12f)
Three multi-node import cycles remained in the Mediator after the 12g
orphan sweep. Each is broken by inverting one edge so dependencies flow
one way, dropping the frozen cycle baseline 5 -> 2.

- ApiDirectCall/Template: move the ICollectionResult type alias to
  RefResolver (where ITemplateScope lives); GenericFingerprintBuilder
  re-exports it. Inverts the builder -> RefResolver back-edge.
- Dashboard: extract resolveAbsoluteHref + helpers into a new leaf
  DashboardHref; DashboardNavigation imports the leaf directly while
  DashboardDiscovery re-exports for external consumers.
- Elements: extract the 3 shared option interfaces into a new leaf
  ElementsActionTypes; relocate capturePageText to its sole consumer
  ElementWaitAction. PageEvalAction imports the leaf, and
  ElementsInteractions re-exports both for existing consumers.

Public surface is unchanged (all relocated symbols stay re-exported).
The 3 relocated Elements interfaces are allowlisted for docs-coverage
(framework-internal, already-public types).
